LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-activation/activation-replace-cross-origin-expected.txt



FAIL navigation.activation after replace cross-origin assert_equals: expected 1 but got 0
